+++ packages/hydrogen/src/types.ts
@@ -131,4 +131,5 @@ export interface CacheOptions {
 
 export interface HydrogenVitePluginOptions {
   devCache?: boolean;
+  purgeQueryCacheOnBuild?: boolean;
 }