+++ packages/cli/src/commands/check/rules/index.ts
@@ -1,0 +1,4 @@
+export {checkHydrogenVersion} from './hydrogen';
+export {checkNodeVersion} from './node';
+export {checkShopify} from './shopify';
+export {checkEslintConfig} from './eslint';