+++ packages/hydrogen/CHANGELOG.md
@@ -13,6 +13,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## 0.8.2 - 2022-01-07
 
 - fix: load logger only once.
+- fix: warn when requests take longer than 3000ms instead of erroring
 - feat: `useQuery` returns an error if the query's fetch was unsuccessful
 - feat: `useShopQuery` will give error hints to look at `shopify.config.js` when the Storefront API responds with a 403
 - fix: do not attempt to decode product IDs, as they are no longer base64-encoded in `unstable`