+++ packages/hydrogen/CHANGELOG.md
@@ -9,6 +9,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - fix: replace log abbreviations with full text.
 - feature: add optional `locale` param to `useShopQuery` to be used as `Accept-Language` in the store Storefront API query
+- feature: Optional purge query cache per build
 
 ## 0.8.2 - 2022-01-07