+++ packages/hydrogen/CHANGELOG.md
@@ -13,6 +13,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 - feature: add optional `locale` param to `useShopQuery` to be used as `Accept-Language` in the store Storefront API query
 - feature: API routes 🎉
 - feature: move to undici instead of node-fetch
+- feature: Optional purge query cache per build
 
 ## 0.8.2 - 2022-01-07